Fix usb boot issue for the case where usb does not appear as sdb.
We had a bug where we were unable to boot a device where the usb
drive appeared as sdc. This is because it was using our fixed set
of devices needed to start X rather than re-using the devices
that are created in the initrd. This fixes that to re-use the
devices created by udev if we have an initramfs.

This CL also fixes a reboot bug.

Review URL: http://chromereview.prom.corp.google.com/1183002


git-svn-id: svn://chrome-svn/chromeos/trunk@174 06c00378-0e64-4dae-be16-12b19f9950a1