Install: tolerate abort.
This problem can happen with the installer:
- User boots from usb; kernel commandline says root=LABEL=C-ROOT
- User installs, which copies the root fs over
- During copy, user aborts (e.g., hits power button)

- User boots from usb; kernel commandline says root=LABEL=C-ROOT
  again. This time, the kernel finds the semi-copied filesystem with
  that label, tries and fails to boot. User is bricked until they get
  a new USB key with a different root label.

Already, after a copy succeeds, the label on the hard drive is
changed. The problem only comes up when the copy is aborted.

The fix here is to copy everything but the first 2K of the root
filesystem over first. Then, copy the first 2K aside and update the
label (which is there), then copy that into the final partition. This
eliminates time when aborting install could lead to a problematic
state.

Credit for the idea goes to Randall Spangler.

temp

Review URL: http://chromereview.prom.corp.google.com/1179001

git-svn-id: svn://chrome-svn/chromeos/trunk@175 06c00378-0e64-4dae-be16-12b19f9950a1