Implementing log out.
This relies on being the direct parent of the chrome browser process, so taht it can kill it and then appropriately wait on it.  If there's something standing between that doesn't wait for its children to die before exiting, then the session manager will still kill the browser (and, indeed, all its descendents) but won't be able to wait for them properly.

We also exit under some new conditions:
1) the browser exiting cleanly
2) receiving a StopSession message over DBus
3) receiving TERM, INT or HUP signals.

Review URL: http://codereview.chromium.org/579002