X is not called Xorg anymore, it seems
One problem seemed to be that, if the session manager exited, it'd be unable to kill X appropriately and thus hang when it tried to start X again.  This fixes that.  Another CL (in a different repo) on the way to stop the infinite restarting problem.

Review URL: http://codereview.chromium.org/661222