Add new GPT partitions, hopefully for the last time.
This change puts these GPT partitions on the image:

1 - stateful partiton
2 - kernel A
3 - rootfs A
4 - kernel B
5 - rootfs B
6 - kernel C (512-byte partition only, nothing in it)
7 - rootfs C (512-byte partition only, nothing in it)
8 - OEM customization (16M)
9 - future use
10 - future use
11 - future use
12 - EFI System Partition (only used by EFI BIOS systems)

Whether creating a USB key or a full install, all partitions are created and
numbered the same way. For the USB key, the B kernel and B rootfs is stubbed
out, so the USB key can't be upgraded via memento (that's nothing new).
/usr/sbin/chromeos-install will duplicate the A image to populate the B
image on the hard drive.

This change will also align active partitions to a 2M boundary to speed up
various development and upgrade functions.

Review URL: http://codereview.chromium.org/1539036