Modify GUIDs to work around a bug in the gpt tool.
Our gpt tool has a bug in the way it writes EFI UUIDs to disk. This change
is a workaround so that the correct GUIDs end up on the disk for the BIOS to
read. The ESP_GUID value was already modified, so it doesn't need changing
at this time.

Review URL: http://codereview.chromium.org/1642014