&& and || do not short-circuit as much as I thought they would, given that they have the same precedence.
false || true || true && echo hi

This echoes "hi" instead of stopping after the first true.

To clarify: the manual does say that && and || have the same precedence (it calls it "priority"), but says nothing about the associativity.  I thought it would be right-associative, but it is left-associative.  Fair enough.

Review URL: http://codereview.chromium.org/1934005