Remove chrome singleton files left from prior run.
If the system is rebooted by force Chrome doesn't get a chance to clean these files up. When Chrome is restarted if any of its initial pids is the same as that of the browser process on the prior run there is a 20 second timeout as the new browser process tries to contact the no longer existing prior one, through its SingletonSocket. This happens more often on cros because everything is always run in the same order (but not every time because there's some timing related differences).

BUG=2886
TEST=Boot system. On command line "sudo reboot". Do this many times. There should never be a 20 second pause before the login panel appears.

Review URL: http://codereview.chromium.org/2015006