chromeos_startup: support device-mapper based rootfs
The pending changes for boot-to-device-mapper and the verified
boot kernel module (dm-verity) change what rootdev sees.  This
change assumes that if rootdev fails (returns empty), we're
using the device mapper and attempts to grab the real root dev
from the sysfs entry.

BUG=none
TEST=manual: gmerged a patched kernel, created a hashed disk image by
patching the verity cl into my build, dd'd that on to the end of /dev/sda3
and configured extlinux.conf on /dev/sda5 to boot to root=/dev/dm-0 with
the given hash. Also tested that normal boot is unimpaired.

Review URL: http://codereview.chromium.org/2114016