Start ibus-daemon and ibus-memconf before Chrome starts to allow Chrome login screen to do the following:
- Get the list of keyboard layouts from ibus-daemon that are necessary for the current UI locale. For example, if current locale is en-US, US qwerty and US dvorak keyboard layouts should be available on the login screen.
- Ask ibus-daemon to change the current keyboard layout. For example, US dvorak should be selected by default if a user selected US dvorak keyboard layout on the login screen last time.

Added ibus-candidate-window.conf so candidate_window process is started on login-prompt-ready event (i.e. after Chrome is started) not to make the power-on-to-login-prompt time slow. Since candidate_window process is not so small, it may take some time to start up.

On the other hand, since ibus-daemon and ibus-memconf are small enough, I believe the impact for the start up time is negligible.

BUG=chromium-os:2619
BUG=chromium-os:2628

Review URL: http://codereview.chromium.org/2875003