Added a DBus signal to be sent when a data connection is established or torn down.
This is to allow flimflam to properly track connection state when the network reports a change.

Get the data bearer technology as early as possible (following an
Enable), so that the RoamingIndicatorCallback never inappropriately
reports a value of zero, which is interpreted by flimflam as meaning
the modem is not registered on a network.

Replaced the SDKCALL macros with ENSURE_SDK_SUCCESS, so that code
that makes SDK calls looks like valid C++ and doesn't obscure what's
going on.

Change "carrier" property returned by GetStatus to "operator_name".

BUG=4559
TEST=Ran cromo with flimflamd. Did connect-service. Then did mm-disconnect (which talks to cromo directly), and observed that flimflamd saw that a disconnect had occurred and set its state accordingly. Did mm-connect and observed that flimflamd handled the connect properly.

Review URL: https://chromereview.prom.corp.google.com/1457010