chromeos-postinst,setimage: enable vboot updates (x86 only)
This changes resolves a number of issues:
- updating efi and legacy bootloader configuration updates for dm-verity arguments
- Updates the rootfs hashes so that we don't ever have to download them.
   (Current tests on similar hardware shows it takes about 8s)
- Autodetects if the USB rootfs if verified and defaults to a verified rootfs on install

The change fixes the source of the syslinux cfg to use the templates in INSTALL_ROOT.
In addition, it adds configuration extraction from the kernel partition to update the
x86 bootloaders.  It also ensures that we don't rewrite bits in the fs superblock by
changing the label.

TEST=build_image --enable_rootfs_verification; install to usb; boot chromeos-vusb.A; install then boot -usb.A and install
BUG=chromium-os:2963

Review URL: http://codereview.chromium.org/3035014

Change-Id: I8d161ef73127b95f9927b6366cd10e60c7d10f8c