Add ambient light sensor support to the power_manager's backlight logic.
Hysteresis of ALS readings is added to BackingController.  I currently set this to 5%, but granularity of 2% in AmbientLightSensor gives 6% hysteresis.

The BacklightController knows about LightSensor via its set_light_sensor() method.  So it can send power events to AmbientLightSensor to enable/disable polling.

The BacklightController is passed to the AmbientLightSensor constructor so it can receive ALS events from it.

This version of AmbientLightSensor is rather tsl2563-specific.  We do not currently support anything else.

BUG=chromium-os:858
: ambient light sensor support
TEST=manually tested on a ZGA board with ambient light sensor and backlight on tigerpoint and pineview instead of the EC.

Three related issues here are 3041003, 2845045, and 2856041.

Review URL: http://codereview.chromium.org/2856041