chromeos-setimage,postinst: add --install_root for verity/dump_kernel_config
verity and dump_kernel_config may be updated in the future. If they are, on
update, the _older_ version will be used.  If the kernel partition format or
dm-verity data formats change, these will be out of sync and break the update.
Or we'll have to account for this by supporting a transition period. This is
easier.  Did I miss any spots for this?

TEST=rev ebuild; build packages; build_image; usb
BUG=none

Change-Id: I270a245d0a20623de9c64ce37d7f4de545da4516

Review URL: http://codereview.chromium.org/3044020