[login_manager] Class to handle loading and setting the owner's public key
If there is an owner key on disk, we will load that key, and deny
attempts to set a new key programmatically.  If there is no key
present, we will allow the owner's key to be set programmatically,
and will persist it to disk upon request.  Attempts to set the key
before on-disk storage has been checked will be denied.

BUG=4486
TEST=unit tests

Change-Id: I04e9e951a35498cb3a9bbe2629be0f09ea3f738f

Review URL: http://codereview.chromium.org/3164015