Propagate low-level TPM errors so that we can better recover from errors.
This is a quick-fix to cryptohome so that it detects some lower-level TPM errors as transient and does not re-create a user's home directory when it encounters them.  Specifically, a communications error, invalid handle error, or defend lock mode error are all now considered transient.  This should keep cryptohome from having too much of an issue if the TPM loses power (for example, if it loses power when it goes to sleep).

Change-Id: I68f2aaa41ec22cff4acb37b0f7808ab157fd5169

BUG=NONE
TEST=Manual

Review URL: http://codereview.chromium.org/3116044