Add asynchronous DBUS APIs for Mount, MountGuest, TestCredentials, and MigrateKey.
This CL adds asynchronous DBUS APIs for methods that may be slow due to crypto operations.  The asynchronous APIs post a task to a worker thread in cryptohome.  Upon completion, a DBUS signal is emitted with the status result.  The change also moves the synchronous versions of these operations to the worker thread as well; the synchronous APIs now merely wait on a completion event before returning their status.  To ensure serialized messages to DBUS, the signal emitted after task completion is marshaled to the main loop.

Change-Id: I8de5aa356226d7e996f87130ecfc8e9dda507b00
Change-Id: I9020d290ea01a16ea6682ed0c7fd4c4976f7a7c3
Change-Id: I4d6e991216ead3218a44330d8dcfe899af89952d

BUG=5447
TEST=unit, manual

Review URL: http://codereview.chromium.org/3146035