Remove legacy process termination code.
This CL removes (very) legacy code that terminates any processes with files open on the user's cryptohome when the call to Unmount() is made.  Generally, the code should not have been executed anymore, so it shouldn't be there.  The assumption is that at the point where Unmount() is called, all processes with open files should have exited.  If not, even with this change, where cryptohomed will not kill the processes, there is still the possibility of failed writes after the lazy unmount because we have to clear the user's kernel keyring (this may mean that subsequent writes may fail, but I haven't verified this in the ecryptfs code).

Change-Id: I4f97370e1c4cf5ef634161c7ac01149f575dbb4d

BUG=NONE
TEST=Manual login, logout

Review URL: http://codereview.chromium.org/3327003