AU: when making a delta update, don't delta compress kernel partition.
This is an unfortunate and hopefully temporarly measure to address a
problem with the images people have in the field. People on dogfood
devices (with traditional BIOS) may have a USB-signed or a HD-signed
kernel partition. This was seen as okay since traditional BIOS doesn't
use the kernel partition. However, this is a problem for delta
updates, as when a delta from usb->usb kernel partitions is applied to
an hd-kernel partition, the resultant kernel partition is
invalid. While that may not seem like a problem (again, since
traditional BIOS doesn't use it to boot), the postinst script at the
end of update DOES look at the new kernel partition and bails if it's
malformed, thus causing the update to fail.

This CL causes delta images to send the kernel partition down in full
(bzip2 compressed), which solves the immediate problem. The rootfs is
still delta compressed.

Longer term, we need to find a way to know exactly what's installed on
the user's hard drive, including kernel partition, so that we can
create a delta from it.

BUG=chromium-os:6390
TEST=Tested update to a dogfood device with traditional BIOS.

Review URL: http://codereview.chromium.org/3325009