Provide an immutable SAFT startup script.
SAFT requires specific startup code running after every
reboot, it used to be remounting root fs as writable and
installing a start up script just for that.

With the rootfs verification enabled it became impossible.

The solution is to introduce a startup script which is
always there and checks if a actual SAFT start script
exists on the stateful partition. If it does, this startup
script invokes it passing along the flash device name.

Change-Id: I8e31c12fb58f0fd1558dc69616b9d3f0c5ab0fdb

BUG=chromium-os:6345
TEST=manual

Tested in concert with http://codereview.chromium.org/3297002 (still being
prepared for review)

Review URL: http://codereview.chromium.org/3330005