cromo: Synchronous shutdown.
Some carriers require that we disconnect cleanly from their network; this change (in combination with 3324005 and 3355006) causes the system to stay up until cromo has disconnected fully. This procedure will also time out after 10 seconds in case of slow or nonexistent uplinks.

To do this, we introduce two tables of hooks which are called a) when we begin shutting down, and b) to ask if we're done shutting down. The pre-exit hooks normally initiate a clean disconnect, with the exit-ok hooks allowing an exit if their device is in a disconnected state. The exit hooks are polled (ugh) at some short interval after the pre-exit hooks are called.

Note that because of the unpleasantness of interacting with glib in a signal-safe way, the signal-handling code here is *exceptionally* ugly - we block signals for all threads except a dedicated target thread, which continuously waits (synchronously) for signals to arrive and dispatches on them. I apologize for this hack; the alternatives were all worse.

There are corresponding changes to all plugins.

BUG=chromium-os:6392
TEST=None
It's difficult to test this without adding a debugging print inside cromo, and
even then, it's hard to tell whether we're /actually/ disconnected or not.

Signed-Off-By: Elly Jones <ellyjones@chromium.org>

Review URL: http://codereview.chromium.org/3291016

Change-Id: I8ab19526c943f220e85bef50c5b2e3308b27e945