Address TPM state regression in offline login.
The fix for TPM state handling introduced a regression in offline login that may result in loss of cryptohome.  This CL fixes it.

The bug is that the error handling in crypto.cc is too liberal, considering all errors but transient communications failures to be fatal.  However, there are two cases where this is incorrect:  first, on resume from S3, we may get TCS_E_KM_LOADFAILED from the TSS API due to state inconsistencies, and second, in offline login, an incorrect password can trigger this (note that screen lock uses a different path).  The fix is to correct the default case to not be fatal and to add the TCS_E_KM_LOADFAILED error as a communications failure with the TPM.

Tests:
1. S3 test.  Pre-patch, this will fail on the final log in, and re-create your cryptohome.  Post-patch, it will fail the final log in (as it detects the error), but a second log in attempt will re-connect to the TPM and succeed (without cryptohome recreation)
  a. Ensure the TPM is owned
  b. Log in
  c. Sleep the system
  d. Unlock the screen
  e. Log out
  f. Log in
2. Offline test.  Pre-patch, (e) will cause cryptohome to be deleted, and (f) will fail because it cannot validate the login.  Post-patch, this will work as expected.
  a. Ensure the TPM is owned
  b. Log in
  c. Log out
  d. Disconnect ethernet, disable wifi
  e. Try to log in with an incorrect password
  f. Try to log in with a correct password

Change-Id: I946deb4005d0265e72b32c81fa847f5d1958f52c

BUG=None
TEST=Attempt login after S3 or incorrect password while offline

Review URL: http://codereview.chromium.org/3355011