gobi-cromo-plugin: Log more events.
We now register for all the callbacks except the CATEvent one, which I a) do not
know the purpose of and b) do not know how to register for in the first place,
since it seems to take two opaque arguments. Other than that, all events are now
handled and logged in the callback threads.

The motivation for this is to make debugging of user problems easier in future; when we're looking at a debug log, we might see suspicious events that could indicate an underlying failure. We currently only log events without a real handler, since all the real handlers do their own logging.

TEST=Adhoc
Run cromo; look at its logfile to see logged events.

BUG=chromium-os:6569

Change-Id: Id324a53762d24b15843efc391b4715f91b02f1f0
Signed-Off-By: Elly Jones <ellyjones@google.com>

Review URL: https://chromereview.prom.corp.google.com/1639016