Add asynchronous version of Remove.
Adds an asychronous version of the Remove() API, and changes the synchronous version to use the worker thread.  This change allows a Remove operation (which might take some time as it must remove all files in a user's vault) to run in the background and signal the caller on completion.

Change-Id: I01cde81825b0d80915d83b470a023983b3268a0c

BUG=None
TEST=Unit

Review URL: http://codereview.chromium.org/3343018