install: make set-image use utilities from the new install root
During an upgrade, we should be using the utilities from the new
install root, not the existing system. This came to a head as we've
found that rootdev on older systems doesn't support "-s".

Also, add --debug option to install,postinst,set-image (and pass it
among them). When set, it will cause "set -x" to be called in each
shell script after flags parsing. This should be useful for debugging.

BUG=chrome-os-partner:1095
TEST=Performed install; nsanders will do factory install before checkin

Review URL: http://codereview.chromium.org/3391005