Explicitly set the size of the vault keyset key.
This CL explicitly sets the size of the vault keyset key to 256-bit (VKK, or local_vault_wrapper in WrapVaultKeyset) instead of implicitly using the size of a passkey (which should always be 256-bit, but isn't explicit).  The CL also sets fallback_to_scrypt_ in crypto.cc to default to true (it is set to true at runtime by mount.cc anyway) so that it is in-line with mount.cc.  Finally, it clarifies the processing in WrapVaultKeyset to describe the TPM/Scrypt flows.

Change-Id: I707c933ff00b173716975c54a7803529cf0ec94b

BUG=None
TEST=Unit tests, manual testing on an image

Review URL: http://codereview.chromium.org/3461004