gobi-cromo-plugin: return an error if OTASP activation fails.
Currently, we can handle delayed failures of OTASP activation (i.e., when we ask
the modem to activate and it responds with "In progress" and later comes back
with an error); however, if the OTASP activation call fails immediately, we
didn't detect that. This fix causes us to return a DBus error when this happens.

TEST=Adhoc
Unfortunately, testing this requires fault injection inside the Qualcomm SDK,
for which no facility currently exists. A small but not redistributable hack to
the SDK will let you simulate an OTASP failure.

BUG=chromium-os:6851

Change-Id: Ic9993b65ab57b155bb645a809bd4d61865198405
Signed-Off-By: Elly Jones <ellyjones@google.com>

Review URL: https://chromereview.prom.corp.google.com/1645008