Add better handling of a TPM clear.
This adds better detection of a TPM clear by enabling an TpmInit callback to detect when ownership is acquired.  When this happens, cryptohome re-establishes its TPM context immediately, instead of passively waiting for the next failed TPM operation.  While this change does not affect normal operation, it does address an issue in the test automation where the TPM was cleared but the stateful partition was not.

In addition, this change adds additional robustness against intermittent TPM failures by adding a one-time retry for Mount APIs that may fail because of a TPM communications failure.

Change-Id: Ib810c8d1627500a4d5f370ba4554e43543c564b9

BUG=6883
TEST=unit, manual

Review URL: http://codereview.chromium.org/3448020