cromo: Add powerd integration.
This adds the ability for cromo to monitor powerd's state using the
NameOwnerChanged DBus signal. It also provides hooks for modules to use to
handle powerd starts and stops (used primarily for registering for suspend
delays, as implemented in 3440007).

TEST=Adhoc
Add test hooks on powerd_up and powerd_down to a plugin of your choice; restart
powerd a couple of times and see that those hooks are called. Try restarting
cromo; notice that the poll for powerd is successful.

BUG=chromium-os:6879

Change-Id: I18347a597a3775a8c24c1175fad10b0a83a1ec8f
Signed-Off-By: Elly Jones <ellyjones@chromium.org>

Review URL: http://codereview.chromium.org/3380023