Move bios logging code out of boot critical path and add switches
This code was moved from chromeos_startup to here to get it off the
critical path for boot. It was recommended to do this in the background
under cover of the X startup. Also report bios switch states.

There is a separate CL to remove the code from the startup script.

BUG=chromium-os:5557,chromium-os:7129
TEST=Build image, check about:system on three different machine types

Change-Id: Ic50a83609c3222999a5d3f0651f70721a7202e01

Review URL: http://codereview.chromium.org/3535002