Add command-line flag to cryptohome to remove tracked dirs.
This adds a command line flag to cryptohome to kick off removal of tracked subdirectories.  It also adds the service-side implementation of the removal.  Finally, it adds some additional diagnostic logging for the (rare) case that Unmount is called when a process still has files open on the mount point.

Change-Id: I10ff042536155991c34d0a98936df623049a3834

BUG=6859
TEST=unit, manual

Review URL: http://codereview.chromium.org/3578006