gobi-cromo-plugin: Add activation timeout support.
In some cases, we can fire off an activation attempt and never hear from the
Gobi again about it, which causes some UI unpleasantness - we have a device, so
'Disable Cellular' appears in the menu (as cellular is present and enabled), but
the device never has a network since it never finishes activating.

TEST=Adhoc
Hack gobi_sdk_wrappers.cc a bit so that ActivateAutomatic always just returns 0,
then activate the modem on a device that hasn't yet been activated.

BUG=chromium-os:6818
We believe this is the root cause of 6818 on some devices; the symptoms of the
behavior being fixed are identical to 6818.

Change-Id: I694f2687705a4be206e65cbe5cfdb560869bd1c8
Signed-off-by: Elly Jones <ellyjones@google.com>
Signed-off-by: Jason Glasgow <jglasgow@google.com>
Signed-off-by: Eric Shienbrood <ers@google.com>

Review URL: https://chromereview.prom.corp.google.com/1661001