cros_boot_mode: make cros_debug bootloader argument prempt others
cros_debug was meant to be a replacement for cros_* to force
developer mode behavior - such as on legacy firmware.  However,
the easiest way to do so without modifying a file is to add the
argument to the bootloader during boot.  With syslinux, it is
appended.  That means that a cros_legacy commandline will have
cros_debug appended.  Since order matters, moving cros_debug
up the chain allows it to preempt any other values.

BUG=chromium-os:824
TEST=unittests

Review URL: http://codereview.chromium.org/3601014

Change-Id: I4cb51216cb9ae9ac9172728da2aba92e75a2411c