Whitelist hotkeys in crosh
This CL introduces an inputrc file used by crosh to disable readline hotkeys that are not part of the whitelist of approved commands.

Unfortunately because of the way bind works, it's not possible (or, more specifically, I spent enough time trying to figure it out that I decided this solution would be Good Enough) to generate this at run time or build time.

BUG=chromium-os:4088
TEST=gmerge onto device

Review URL: http://codereview.chromium.org/3733002

Change-Id: Ic90a34aa7af18dac7e23c26ef0cb8b1df379eb07