[login_manager] Whitelist owner in session_manager
Let the session manager handle whitelisting the owner, and adding the
"device owner" preference -- instead of having Chrome do it.

For max robustness, try at the start of every session.  We'll only be
able to do it successfully if the currently logged-in user is the owner.

BUG=chromium-os:7511
TEST=unit tests, install on device and test.

Change-Id: If04ed631a8a63700f1394a0de50e7f36845b29fe

Review URL: http://codereview.chromium.org/3732003