login_manager: Avoid killing window manager.
This makes us avoid killing jobs with argv[0]s ending in
"window-manager-session.sh" when the session stops, so that
the window manager will keep control of the screen until X
gets killed.

It also cleans up the job-killing code so that we send TERM
to all of the processes and then go back and wait for them
to exit, rather than sending TERM to one, waiting, sending
TERM to the next, etc.

I also noticed that SystemUtils::ChildIsGone() may have
problems if the system clock gets stepped backwards
(unlikely, but I've seen it cause problems in the window
manager before), so I'm updating it to use the monotonically-
increasing clock instead.

BUG=chromium-os:7901
TEST=added

Review URL: http://codereview.chromium.org/3947001

Change-Id: I2bb34646003fb0ab3d758e469bfedbe1fa4a9e9b